+++ CHANGELOG.md
@@ -38,6 +38,8 @@ Translations: [简体中文](CHANGELOG_zh.md)
   used in placeholder and fallback
 * improve: Improve IconDrawable, support fixed-size background and restrict icons to have fixed
   sizes or specify iconSize
+* improve: Improve IconPainter, support fixed-size background and restrict icons to have fixed
+  sizes or specify iconSize
 * new: Non-Android platforms now also support ColorType
 * new: Non-Android platforms now also support ColorSpace
 * new: DrawableDecoder supports colorSpace